# doc/swagger.yml
openapi: 3.0.1
info:
  title: dating server API
  description: Описание методов API сервера мобильного приложения
  version: 0.24.6
servers:
  - url: https://dev.dating.eagledev.ru/api/v1
  - url: https://prod.dating.eagledev.ru/api/v1
  - url: http://localhost:8032/api/v1/
paths:
...
  /admin/auth:
    post:
      tags:
        - Панель управления
      summary: Вход или регистрация
      description: "Авторизация админа"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
        required: true
      responses:
        "200":
          description: успешная авторизация
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: request success
                    example: true
                  data:
                    required:
                      - token
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT-токен для авторизации запросов
                        example: string-jwt-token

        "400":
          description: некорректный формат запроса/ошибка в теле запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: data
  /admin/dashboard:
    get:
      tags:
        - Панель управления
      security:
        - bearerAuth: [ ]
      summary: Статистика пользователей
      description: возвращает статистику пользователей
      responses:
        "200":
          description: ОК
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - success
                  - sexStats
                properties:
                  success:
                    type: boolean
                    description: request success
                    example: true
                  data:
                    $ref: '#/components/schemas/DashBoard'
        "500":
          description: server error
          content: { }
  /admin/users:
    get:
      tags:
        - Панель управления
      summary: Список пользователей
      security:
        - bearerAuth: [ ]
      description: возвращает список пользователей
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: состояние здоровья сервиса
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsersInfo'
        "500":
          description: server error
          content: { }
  /admin/user/{id}:
    get:
      tags:
        - Панель управления
      summary: Информация о пользователе
      security:
        - bearerAuth: [ ]
      description: возвращает информацию о пользователе
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: состояние здоровья сервиса
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserInfo'
        "500":
          description: server error
          content: { }
...
components:
  schemas:
...
    AdminLoginRequest:
      required:
        - password
        - login
      type: object
      properties:
        login:
          type: string
          description: login
          example: clique
        password:
          type: string
          description: password
          example: "123"
    UsersInfo:
      required:
        - ID
        - name
        - phone
        - status
        - photo
      type: object
      properties:
        ID:
          type: string
          description: uuid сообщения
          example: uuid
        name:
          type: string
          description: "name пользователя"
          example: name
        phone:
          type: string
          description: телефон(временно почта)
          example: "example@mail.ru"
        status:
          type: string
          description: статус пользователя
          example: "active"
        photo:
          type: array
          description: фотография
          example: "73href72631f"
    DashBoardSex:
      required:
        - sexStats
        - count
        - male
        - female
        - undefined
      type: object
      properties:
        male:
          type: integer
          description: количество
          example: 1
        female:
          type: integer
          description: количество
          example: 1
        undefined:
          type: integer
          description: количество
          example: 1
    UserInfo:
      required:
        - ID
        - name
        - phone
        - status
        - photo
        - updatedAt
        - createdAt
        - birthday
        - sex
        - about
        - geolocation
      type: object
      properties:
        ID:
          type: string
          description: uuid сообщения
          example: uuid
        name:
          type: string
          description: "name пользователя"
          example: Ivan
        phone:
          type: string
          description: телефон(временно почта)
          example: "example@mail.ru"
        status:
          type: string
          description: статус пользователя
          example: "active"
        photo:
          type: array
          items:
            $ref: '#/components/schemas/UserPhoto'
        updatedAt:
          type: string
          description: Дата обновления в формате YYYY-MM-DD-THH:MM:SS
          example: "2024-03-27T00:00:00Z"
        createdAt:
          type: string
          description: Дата создания в формате YYYY-MM-DD-THH:MM:SS
          example: "2024-03-27T00:00:00Z"
        birthday:
          type: string
          description: Дата рождения в формате YYYY-MM-DD
          example: 2000-06-29
        sex:
          type: string
          description: пол
          example: "male"
        about:
          type: string
          description: обо мне
          example: "ем куриный суп"
        geolocation:
          $ref: '#/components/schemas/Geolocation'
    Geolocation:
      required:
        - lng
        - lat
      type: object
      properties:
        lng:
          type: number
          description: долгота
          example: 1
        lat:
          type: number
          description: широта
          example: 1
    DashBoard:
      required:
        - totalUsers
        - usersByDay
        - usersByWeek
        - usersActive
        - usersNew
        - usersInActive
      type: object
      properties:
        totalUsers:
          type: integer
          description: Количество пользователей всего
          example: 1
        usersByDay:
          type: integer
          description: Количество пользователей зарегистрированных за день
          example: 1
        usersByWeek:
          type: integer
          description: Количество пользователей зарегистрированных за неделю
          example: 1
        usersActive:
          type: integer
          description: Количество активных пользователей
          example: 1
        usersNew:
          type: integer
          description: Количество новых пользователей
          example: 1
        usersInActive:
          type: integer
          description: Количество неактивных пользователей
          example: 1
        sexStats:
          $ref: '#/components/schemas/DashBoardSex'
...
    ErrorResponse:
      required:
        - error
        - success
      type: object
      properties:
        success:
          type: boolean
          description: request success
          example: false
        error:
          required:
            - code
            - message
          type: object
          properties:
            code:
              type: integer
              description: internal error code
              example: 1001
            message:
              type: string
              description: error text
              example: example error
...
    UserPhoto:
      required:
        - ID
        - URL
        - isMain
      type: object
      properties:
        ID:
          type: string
          description: uuid фотографии
          example: 6a7485ca-2020-470d-b0b5-04f3ff19f12f
        URL:
          type: string
          description: URL адрес картинки
          example: https://example.example/example.jpg
        isMain:
          type: boolean
          description: флаг основное фото профиля
          example: false
        createdAt:
          type: string
          description: дата добавления
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT